backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
mysql-connector-python
mysqlclient
authlib
python-jose